+++ contracts/test/levels/Elevator.t.sol
@@ -1,0 +1,63 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.0;
+
+import "forge-std/Test.sol";
+import {Utils} from "test/utils/Utils.sol";
+
+import {Elevator} from "src/levels/Elevator.sol";
+import {ElevatorFactory} from "src/levels/ElevatorFactory.sol";
+import {ElevatorAttack} from "src/attacks/ElevatorAttack.sol";
+import {Level} from "src/levels/base/Level.sol";
+import {Ethernaut} from "src/Ethernaut.sol";
+
+contract TestElevator is Test, Utils {
+    Ethernaut ethernaut;
+    Elevator instance;
+
+    address payable owner;
+    address payable player;
+
+    /*//////////////////////////////////////////////////////////////
+                                 HELPERS
+    //////////////////////////////////////////////////////////////*/
+
+    function setUp() public {
+        address payable[] memory users = createUsers(2);
+
+        owner = users[0];
+        vm.label(owner, "Owner");
+
+        player = users[1];
+        vm.label(player, "Player");
+
+        vm.startPrank(owner);
+        ethernaut = getEthernautWithStatsProxy(owner);
+        ElevatorFactory factory = new ElevatorFactory();
+        ethernaut.registerLevel(Level(address(factory)));
+        vm.stopPrank();
+
+        vm.startPrank(player);
+        instance = Elevator(payable(createLevelInstance(ethernaut, Level(address(factory)), 0)));
+        vm.stopPrank();
+    }
+
+    /*//////////////////////////////////////////////////////////////
+                                 TESTS
+    //////////////////////////////////////////////////////////////*/
+
+    /// @notice Check the initial state of the level and environment.
+    function testInit() public {
+        vm.startPrank(player);
+        assertFalse(submitLevelInstance(ethernaut, address(instance)));
+    }
+
+    /// @notice Test the solution for the level.
+    function testSolve() public {
+        vm.startPrank(player);
+
+        ElevatorAttack attack = new ElevatorAttack();
+        attack.attack(address(instance));
+
+        assertTrue(submitLevelInstance(ethernaut, address(instance)));
+    }
+}